Append fix for content and excerpt helpers
Updated the helpers for content and excerpt to accept an `append` option
that will allow the user to set a custom append string if the
content/excerpt is truncated. This option is stored in the
`truncateOptions` variable and passed to Downsize before rendering.

There is currently a bug in the Downsize package that will not append
the string if the content does not contain HTML. I included a 'hack' in
this to work around this bug until Downsize merges a patch
(cgiffard/Downsize#19).